refactor transformer decoder
