+++ eole/decoders/__init__.py
@@ -1,7 +1,6 @@
 """Module defining decoders."""
 from eole.decoders.rnn_decoder import InputFeedRNNDecoder, StdRNNDecoder
-from eole.decoders.transformer_decoder import TransformerDecoder
-from eole.decoders.transformer_lm_decoder import TransformerLMDecoder
+from eole.decoders.transformer import TransformerDecoder
 from eole.decoders.cnn_decoder import CNNDecoder
 
 
@@ -10,5 +9,4 @@ str2dec = {
     "ifrnn": InputFeedRNNDecoder,
     "cnn": CNNDecoder,
     "transformer": TransformerDecoder,
-    "transformer_lm": TransformerLMDecoder,
 }
