+++ eole/decoders/cnn_decoder.py
@@ -22,6 +22,7 @@ class CNNDecoder(DecoderBase):
         self,
         model_config,
         running_config=None,
+        with_cross_attn=False,
     ):
         super(CNNDecoder, self).__init__()
 
@@ -51,11 +52,12 @@ class CNNDecoder(DecoderBase):
         )
 
     @classmethod
-    def from_config(cls, model_config, running_config=None):
+    def from_config(cls, model_config, running_config=None, with_cross_attn=False):
         """Alternate constructor."""
         return cls(
             model_config,
             running_config,
+            with_cross_attn=False,
         )
 
     def init_state(self, **kwargs):
