+++ eole/decoders/decoder.py
@@ -15,7 +15,7 @@ class DecoderBase(nn.Module):
         self.state = {}
 
     @classmethod
-    def from_config(cls, model_config, running_config=None):
+    def from_config(cls, model_config, running_config=None, with_cross_attn=False):
         """Alternate constructor.
 
         Subclasses should override this method.
