+++ eole/decoders/rnn_decoder.py
@@ -22,6 +22,7 @@ class RNNDecoderBase(DecoderBase):
         self,
         model_config,
         running_config=None,
+        with_cross_attn=False,
     ):
         super(RNNDecoderBase, self).__init__(
             attentional=model_config.global_attention != "none"
@@ -67,12 +68,13 @@ class RNNDecoderBase(DecoderBase):
             )
 
     @classmethod
-    def from_config(cls, model_config, running_config=None):
+    def from_config(cls, model_config, running_config=None, with_cross_attn=False):
         """Alternate constructor."""
         # config = opt.model.decoder  # RnnDecoderConfig
         return cls(
             model_config,
             running_config=running_config,
+            with_cross_attn=False,
         )
 
     def init_state(self, **kwargs):
@@ -189,6 +191,7 @@ class StdRNNDecoder(RNNDecoderBase):
         self,
         model_config,
         running_config=None,
+        with_cross_attn=False,
     ):
         self.hidden_size = model_config.hidden_size
         self._input_size = model_config.tgt_word_vec_size
@@ -268,6 +271,7 @@ class InputFeedRNNDecoder(RNNDecoderBase):
         self,
         model_config,
         running_config=None,
+        with_cross_attn=False,
     ):
         self.hidden_size = model_config.hidden_size
         self._input_size = model_config.tgt_word_vec_size + self.hidden_size
