+++ eole/tests/test_model_lm.yml
@@ -113,8 +113,8 @@ share_vocab: true
 # feat_vec_exponent=0.7,
 # model_type='text',
 # model_dtype='fp32',
-# encoder_type='transformer_lm', 
-# decoder_type='transformer_lm',
+# encoder_type='transformer', 
+# decoder_type='transformer',
 # layers=-1,
 # enc_layers=2,
 # dec_layers=2,
